test(QuestionTest) : 질문 삭제 도메인 메소드 테스트 코드 작성
- 기존 QuestionTest 를 Repository Test 로 변경

- Question 도메인을 위한 테스트 코드 추가
